Tried to do prediction stuff, not great
Wrote some code to continue a time series from the point when it's fired up again with the best known match for that point. However it's slow and has some other issues but it could still be useful for AI training later on (e.g. for expanding heating periods in a way that the AI could learn how it would continue if no one fires up the furnace again. Otherwise I don't know yet how you could tell the AI that it cannot predict an increase in furnace temperature unless the door has opened and stuff like that. There are quite a lot of constraints to incorporate in this AI and with my very limited knowledge I have no clue how yet).